requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
python-dotenv==0.13.0
requests==2.23.0
SQLAlchemy==1.3.16